refactor: extract serverWrapper
